Partitioned wrapper structure for organization
