+++ makeflow/src/makeflow_wrapper_monitor.h
@@ -1,0 +1,33 @@
+
+/*
+Copyright (C) 2015- The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file COPYING for details.
+*/
+
+#ifndef MAKEFLOW_WRAPPER_MONITOR_H
+#define MAKEFLOW_WRAPPER_MONITOR_H
+
+/*
+This module implements garbage collection on the dag.
+Files that are no longer needed as inputs to any rules
+may be removed, according to a variety of criteria.
+*/
+
+struct makeflow_monitor {
+	struct makeflow_wrapper *wrapper;
+	int enable_time_series;
+	int enable_list_files;
+
+	int interval;
+	char *log_prefix;
+	char *limits_name;
+	char *exe;
+	const char *exe_remote;
+};
+
+struct makeflow_monitor * makeflow_monitor_create();
+void makeflow_prepare_for_monitoring( struct makeflow_monitor *m, char *log_dir, char *log_format);
+char *makeflow_wrap_monitor( char *result, struct dag_node *n, struct makeflow_monitor *m );
+
+#endif
